// src/shared/api/generated/model/problemResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Amazing EKB Hub API
 * Amazing EKB backend contract.
 * OpenAPI spec version: 1.0.0
 */
import type { ProblemResponseDtoCode } from './problemResponseDtoCode'

export type ProblemResponseDto = {
  /** @pattern ^[A-Z][A-Z0-9_]*$ */
  code: ProblemResponseDtoCode
  detail: string
  instance: string
  requestId: string
  /**
   * @minimum 400
   * @maximum 599
   */
  status: number
  title: string
  type: string
}